+++ sharding-orchestration/sharding-orchestration-core/src/main/java/io/shardingsphere/orchestration/internal/config/listener/PropertiesOrchestrationListener.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.orchestration.internal.config.listener;
+
+import io.shardingsphere.core.event.ShardingEventBusInstance;
+import io.shardingsphere.orchestration.internal.config.ConfigurationNode;
+import io.shardingsphere.orchestration.internal.config.ConfigurationService;
+import io.shardingsphere.orchestration.internal.event.config.PropertiesChangedEvent;
+import io.shardingsphere.orchestration.internal.listener.OrchestrationListener;
+import io.shardingsphere.orchestration.reg.api.RegistryCenter;
+import io.shardingsphere.orchestration.reg.listener.DataChangedEvent;
+import io.shardingsphere.orchestration.reg.listener.EventListener;
+
+/**
+ * Properties listener manager.
+ *
+ * @author panjuan
+ */
+public final class PropertiesOrchestrationListener implements OrchestrationListener {
+    
+    private final ConfigurationNode configNode;
+    
+    private final RegistryCenter regCenter;
+    
+    private final ConfigurationService configService;
+    
+    public PropertiesOrchestrationListener(final String name, final RegistryCenter regCenter) {
+        configNode = new ConfigurationNode(name);
+        this.regCenter = regCenter;
+        configService = new ConfigurationService(name, regCenter);
+    }
+    
+    @Override
+    public void watch() {
+        regCenter.watch(configNode.getPropsPath(), new EventListener() {
+            
+            @Override
+            public void onChange(final DataChangedEvent event) {
+                if (DataChangedEvent.Type.UPDATED == event.getEventType()) {
+                    ShardingEventBusInstance.getInstance().post(new PropertiesChangedEvent(configService.loadProperties()));
+                }
+            }
+        });
+    }
+}
